Move {ret,add} from Builder to BasicBlock
