+++ src/ir/builder.rs
@@ -1,7 +1,6 @@
 use std::vec::Vec;
 use super::block::BasicBlock;
 use super::instruction::Instruction;
-use super::register::Register;
 use super::value::Value;
 
 pub struct Builder(Vec<BasicBlock>);
@@ -20,16 +19,6 @@ impl Builder {
         self.0.push(bb);
         label
     }
-    pub fn ret(&mut self, val: Value) {
-        let ret = Instruction::Ret(val);
-        self.push(ret)
-    }
-    pub fn add(&mut self, lhs: Value, rhs: Value) -> Value {
-        let reg = Register::new();
-        let add = Instruction::Add(reg.clone(), lhs, rhs);
-        self.push(add);
-        Value::Register(reg)
-    }
 }
 
 #[cfg(test)]
@@ -39,9 +28,5 @@ mod tests {
     fn test_builder() {
         let mut builder = Builder::new();
         let _bl = builder.block();
-        let lhs = Value::Constant(0);
-        let rhs = Value::Constant(1);
-        let add = builder.add(lhs, rhs);
-        builder.ret(add);
     }
 }
